ctdb-daemon: Remove helper variable CTDB_CLUSTER_MUTEX_HELPER
Use path_helperdir() to help construct the path and then cache the
result in the existing static buffer (with length adjusted because
POSIX says the +1 is not necessary).  Given the way this is used by
cluster_mutex_test, there is no (other) sane place to cache it.
path_helperdir_append() could be used to construct the path, but then
there would be an unnecessary talloc() result to free.

The flexibility in unit test cluster_mutex_003.sh was never used, so
remove this test.  If other cluster mutex helpers are added then they
can be tested by separate tests.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Reviewed-by: Anoop C S <anoopcs@samba.org>